changed float to double for biasEn and control variables due to compiler compatibility issues with Linux. (same goes for TID_FLOAT to TID_DOUBLE) for these variables)
When you run this on the fgddaq@ladd05, delete the Variables and Settings folder in the folder ASUM to re-create all the variables with the new correct types again.

-Brian Lee

SVN revision: 3765

+++ drivers/device/dd_mscbasum.c
@@ -338,7 +338,7 @@ INT mscbasum(INT cmd, ...)
    case CMD_SET_VBIAS:  // Voltage
       info = va_arg(argptr, void *);
       channel = va_arg(argptr, INT);
-      value = (float) va_arg(argptr, double);
+      value = (float) va_arg(argptr, float);
     printf("Set Bias Voltage %d :[%i]%f\n", (indexNum-4), channel, value);
     status = mscbasum_setDacBias(info, channel, value);
       //ss_sleep(10);
@@ -354,7 +354,7 @@ INT mscbasum(INT cmd, ...)
    case CMD_SET_CONTROL:  // Update control variable
       info = va_arg(argptr, void *);
       channel = va_arg(argptr, INT);
-      value = (float) va_arg(argptr, double);
+      value = (float) va_arg(argptr, float);
       status = mscbasum_controlUpdate(info, channel, value);
       break;
 
@@ -368,14 +368,14 @@ INT mscbasum(INT cmd, ...)
   case CMD_SET_BIAS_EN:  // Update Bias Enable variable
     info = va_arg(argptr, void *);
     channel = va_arg(argptr, INT);
-    value = (float) va_arg(argptr, double);
+    value = (float) va_arg(argptr, float);
     status = mscbasum_biasEnUpdate(info, channel, value);
     break;
 
   case CMD_SET_ASUMDACTH:
     info = va_arg(argptr, void *);
       channel = va_arg(argptr, INT);
-      value = (float) va_arg(argptr, double);
+      value = (float) va_arg(argptr, float);
       printf("Set ASUM DAC Threshold voltage:[%i]%f\n", channel, value);
       status = mscgasum_setasumDacTh(info, channel, value);
       break;
@@ -397,7 +397,7 @@ INT mscbasum(INT cmd, ...)
   case CMD_SET_CHPUMPDAC:
   info = va_arg(argptr, void *);
    channel = va_arg(argptr, INT);
-   value = (float) va_arg(argptr, double);
+   value = (float) va_arg(argptr, float);
    printf("Set Charge Pump DAC voltage:[%i]%f\n", channel, value);
    status = mscgasum_setChPumpDac(info, channel, value);
    break;
@@ -420,7 +420,7 @@ INT mscbasum(INT cmd, ...)
    case CMD_SET_CURRENT_LIMIT:
       info = va_arg(argptr, void *);
       channel = va_arg(argptr, INT);
-      value = (float) va_arg(argptr, double);
+      value = (float) va_arg(argptr, float);
       status = mscbasum_setDacBias_current_limit(info, channel, value);
       break;
